updated app.jsx, updated gitignore
