+++ react-rpoject/.gitignore
@@ -4,6 +4,7 @@
 /node_modules
 /.pnp
 .pnp.js
+/src/firebaseConfig.js
 
 # testing
 /coverage
