--- react-rpoject/src/App.js
@@ -1,51 +1,0 @@
-import React from 'react';
-import withFirebaseAuth from 'react-with-firebase-auth'
-import * as firebase from 'firebase/app';
-import 'firebase/auth';
-import firebaseConfig from './firebaseConfig';
-
-const firebaseApp = firebase.initializeApp(firebaseConfig);
-
-const firebaseAppAuth = firebaseApp.auth();
-const providers = {
-  googleProvider: new firebase.auth.GoogleAuthProvider(),
-};
-
-
-class signIn extends React.Component{
-  render(){
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          {
-            user 
-              ? <p>Hello, {user.displayName}</p>
-              : <p>Please sign in.</p>
-          }
-          {
-            user
-              ? <button onClick={signOut}>Sign out</button>
-              : <button onClick={signInWithGoogle}>Sign in with Google</button>
-          }
-        </header>
-      </div>
-    );
-
-  }
-};
-
-const title = () => {
-  return (
-    <div className="title">
-        <h1> React Quiz</h1>
-    </div>
-  )
-}
-
-export {signIn, title, withFirebaseAuth({
-  providers,
-  firebaseAppAuth,
-})(signIn)};
-
-
