+++ react-rpoject/src/index.js
@@ -4,6 +4,4 @@ import ReactDOM from 'react-dom';
 import App from './App';
 
 
-ReactDOM.render(<App />, document.getElementById('root'));
-
-
+ReactDOM.render( < App / > , document.getElementById('root'));
